chore(tests): add validateFolderName, cleanup and move tests
